LayoutTests/platform/glib/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-request-resources.https-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test1.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test2.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test3.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test4.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test5.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test6.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test7.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test8.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test9.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test10.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test11.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test12.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test13.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test14.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test15.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test16.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test17.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test18.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test19.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test20.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test21.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test22.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test23.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test24.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test25.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test26.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test27.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test28.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test29.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test30.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test31.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test32.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test33.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test34.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test35.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test36.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test37.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test38.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test39.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test40.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test41.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test42.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test43.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test44.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test45.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test46.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test47.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test48.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test49.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Fetch API cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test50.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test51.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test52.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test53.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test54.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test55.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test56.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test57.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test58.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/sample?test59.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test60.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test61.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test62.

PASS Verify FetchEvent for resources.
PASS Image load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test1 cross_origin:)
PASS Image load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test2 cross_origin:)
PASS CSS load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test3 cross_origin:)
PASS CSS load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test4 cross_origin:)
PASS Image load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test5 cross_origin:anonymous)
PASS Image load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test6 cross_origin:use-credentials)
PASS Image load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test7 cross_origin:)
PASS Image load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test8 cross_origin:anonymous)
PASS Image load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test9 cross_origin:use-credentials)
PASS Script load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test10 cross_origin:)
PASS Script load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test11 cross_origin:anonymous)
PASS Script load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test12 cross_origin:use-credentials)
PASS Script load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test13 cross_origin:)
PASS Script load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test14 cross_origin:anonymous)
PASS Script load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test15 cross_origin:use-credentials)
PASS CSS load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test16 cross_origin:)
PASS CSS load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test17 cross_origin:anonymous)
PASS CSS load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test18 cross_origin:use-credentials)
PASS CSS load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test19 cross_origin:)
PASS CSS load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test20 cross_origin:anonymous)
PASS CSS load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test21 cross_origin:use-credentials)
FAIL FontFace load (url: https://web-platform.test:9443/service-workers/service-worker/resources/sample?test22) assert_equals: mode of must be cors. expected "cors" but got "no-cors"
FAIL FontFace load (url: https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test23) assert_equals: mode of must be cors. expected "cors" but got "no-cors"
PASS Script load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test24)
PASS Script load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test25)
PASS Script load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test26)
PASS Script load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test27)
PASS Script load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test28)
PASS Script load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test29)
PASS CSS load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test30)
PASS CSS load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test31)
PASS CSS load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test32)
PASS CSS load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test33)
PASS CSS load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test34)
PASS CSS load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test35)
PASS fetch (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test36 mode:same-origin credentials:omit)
PASS fetch (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test37 mode:same-origin credentials:same-origin)
PASS fetch (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test38 mode:same-origin credentials:include)
PASS fetch (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test39 mode:no-cors credentials:omit)
PASS fetch (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test40 mode:no-cors credentials:same-origin)
PASS fetch (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test41 mode:no-cors credentials:include)
PASS fetch (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test42 mode:cors credentials:omit)
PASS fetch (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test43 mode:cors credentials:same-origin)
PASS fetch (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test44 mode:cors credentials:include)
PASS fetch (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test45 mode:no-cors credentials:omit)
PASS fetch (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test46 mode:no-cors credentials:same-origin)
PASS fetch (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test47 mode:no-cors credentials:include)
PASS fetch (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test48 mode:cors credentials:omit)
PASS fetch (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test49 mode:cors credentials:same-origin)
PASS fetch (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test50 mode:cors credentials:include)
PASS Audio load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test51 cross_origin:)
PASS Audio load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test52 cross_origin:anonymous)
PASS Audio load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test53 cross_origin:use-credentials)
PASS Audio load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test54 cross_origin:)
PASS Audio load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test55 cross_origin:anonymous)
PASS Audio load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test56 cross_origin:use-credentials)
PASS Video load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test57 cross_origin:)
PASS Video load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test58 cross_origin:anonymous)
PASS Video load (url:https://web-platform.test:9443/service-workers/service-worker/resources/sample?test59 cross_origin:use-credentials)
PASS Video load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test60 cross_origin:)
PASS Video load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test61 cross_origin:anonymous)
PASS Video load (url:https://www1.web-platform.test:9443/service-workers/service-worker/resources/sample?test62 cross_origin:use-credentials)
